BUILD: Resolve some build warnings around bundling
